requirements.txt
yara-python
elftools
argparse